+++ test/successful-files.test.js
@@ -363,3 +363,29 @@ describe('a custom loader entry can include only a sync loader and work for both
     checkResult(result);
   });
 });
+
+describe('works fine if sync loader returns a Promise from a JS file', () => {
+  beforeEach(() => {
+    temp.createFile(
+      'foo.config.js',
+      'module.exports = Promise.resolve({ a: 1 });'
+    );
+  });
+
+  const file = temp.absolutePath('foo.config.js');
+  const explorerOptions = {
+    stopDir: temp.absolutePath('.'),
+    searchPlaces: ['foo.config.js'],
+  };
+
+  test('sync', () => {
+    const result = cosmiconfig('foo', explorerOptions).loadSync(file);
+    expect(result).toEqual({
+      filepath: file,
+      config: expect.any(Promise),
+    });
+    return result.config.then(resolved => {
+      expect(resolved).toEqual({ a: 1 });
+    });
+  });
+});